SE-299 Removed dependency and cleaned up code
