--- sdk/features.yml
@@ -1,62 +1,0 @@
-LusidExamples:
- - SDKs:
-  - feature: F1
-    description: "Apply Bi-temporal portfolio change"
-  - feature: F2
-    description: "Get holdings"
-  - feature: F3
-    description: "Set target holdings"
-  - feature: F4
-    description: "Upsert simple order"
-  - feature: F5
-    description: "Upsert simple order with unknown instrument"
-  - feature: F6
-    description: "Update simple order"
-  - feature: F7
-    description: "Upsert and retrieve simple orders"
-  - feature: F8
-    description: "Create transaction portfolio"
-  - feature: F9
-    description: "Create transaction portfolio with property"
-  - feature: F10
-    description: "Add transactions to portfolio"
-  - feature: F11
-    description: "Add transactions to portfolio with property"
-  - feature: F12
-    description: "List scopes"
-  - feature: F13
-    description: "List portfolios"
-  - feature: F14
-    description: "Create portfolio with label property"
-  - feature: F15
-    description: "Create portfolio with metric property"
-  - feature: F16
-    description: "Reconcile portfolio"
-  - feature: F17
-    description: "Load listed instrument transaction"
-  - feature: F18
-    description: "Load cash transaction"
-  - feature: F19
-    description: "Load otc instrument transaction"
-  - feature: F20
-    description: "Run aggregation with buy"
-  - feature: F21
-    description: "Run aggregation on multicurrency portfolio with stock in"
-  - feature: F22
-    description: "Lookup instrument by unique id"
-  - feature: F23
-    description: "List available identifiers"
-  - feature: F24
-    description: "List all instruments"
-  - feature: F25
-    description: "List instruments by identifier type"
-  - feature: F26
-    description: "Edit instrument property"
-  - feature: F27
-    description: "Create custom instrument"
-  - feature: F28
-    description: "Add quote"
-  - feature: F29
-    description: "Add quote for instrument for single day"
-  - feature: F30
-    description: "Get timeseries quote"
