+++ sdk/pom.xml
@@ -228,7 +228,23 @@
 
     </profiles>
 
+    <repositories>
+        <repository>
+            <id>osssrh</id>
+            <name>OSSRH</name>
+            <url>https://oss.sonatype.org/content/repositories/snapshots/</url>
+            <snapshots>
+                <enabled>true</enabled>
+            </snapshots>
+        </repository>
+    </repositories>
+
     <dependencies>
+        <dependency>
+            <groupId>com.finbourne</groupId>
+            <artifactId>lusid-features</artifactId>
+            <version>1.0.1-SNAPSHOT</version>
+        </dependency>
         <dependency>
             <groupId>io.swagger</groupId>
             <artifactId>swagger-annotations</artifactId>
@@ -292,11 +308,6 @@
             <artifactId>mockito-core</artifactId>
             <version>${mockito-version}</version>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.reflections</groupId>
-            <artifactId>reflections</artifactId>
-            <version>0.9.9-RC1</version>
         </dependency>
     </dependencies>
