+++ sdk/src/test/tutorial/java/com/finbourne/lusid/tutorials/ibor/parity/FeatureLookthrough.java
@@ -1,39 +1,96 @@
 package com.finbourne.lusid.tutorials.ibor.parity;
 
-import com.finbourne.lusid.Configuration;
-import org.reflections.Reflections;
-import org.reflections.scanners.MethodAnnotationsScanner;
-import org.reflections.util.ClasspathHelper;
-import org.reflections.util.ConfigurationBuilder;
-
-import java.lang.annotation.Annotation;
+import java.io.File;
+import java.io.IOException;
 import java.lang.reflect.Method;
-import java.util.Set;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Enumeration;
+import java.util.List;
 
 public class FeatureLookthrough {
 
-    @LusidFeature(value = "Test")
-    public void retrieveAnnotations() {
-        ConfigurationBuilder configurationBuilder = new ConfigurationBuilder().setUrls(
-                ClasspathHelper.forPackage("com.finbourne.lusid.tutorials")
-        ).setScanners(new MethodAnnotationsScanner());
+    String annotations;
 
-        Reflections reflections = new Reflections(configurationBuilder);
-
-        Set<Method> methods = reflections.getMethodsAnnotatedWith(LusidFeature.class);
-        System.out.println(ClasspathHelper.forPackage("com.finbourne.lusid.tutorials"));
-
-
-        for (Method method : methods) {
-//            if (method.isAnnotationPresent(LusidFeature.class))
-//            {
-            // do something
-            System.out.println(method.getAnnotation(LusidFeature.class));
-            System.out.println(method.isAnnotationPresent(LusidFeature.class));
-//            }
+    /**
+     * Scans all classes accessible from the context class loader which belong
+     * to the given package and subpackages.
+     *
+     * @param packageName
+     *            The base package
+     * @return The classes
+     * @throws ClassNotFoundException
+     * @throws IOException
+     */
+    private Iterable<Class> getClasses(String packageName) throws ClassNotFoundException, IOException, URISyntaxException {
+        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
+        String path = packageName.replace('.', '/');
+        Enumeration<URL> resources = classLoader.getResources(path);
+        List<File> dirs = new ArrayList<File>();
+        while (resources.hasMoreElements())
+        {
+            URL resource = resources.nextElement();
+            URI uri = new URI(resource.toString());
+            dirs.add(new File(uri.getPath()));
+        }
+        List<Class> classes = new ArrayList<Class>();
+        for (File directory : dirs)
+        {
+            classes.addAll(findClasses(directory, packageName));
         }
 
+        return classes;
     }
 
+    /**
+     * Recursive method used to find all classes in a given directory and
+     * subdirs.
+     *
+     * @param directory
+     *            The base directory
+     * @param packageName
+     *            The package name for classes found inside the base directory
+     * @return The classes
+     * @throws ClassNotFoundException
+     */
+    private List<Class> findClasses(File directory, String packageName) throws ClassNotFoundException
+    {
+        List<Class> classes = new ArrayList<Class>();
+        if (!directory.exists())
+        {
+            return classes;
+        }
+        File[] files = directory.listFiles();
+        for (File file : files)
+        {
+            if (file.isDirectory())
+            {
+                classes.addAll(findClasses(file, packageName + "." + file.getName()));
+            }
+            else if (file.getName().endsWith(".class"))
+            {
+                classes.add(Class.forName(packageName + '.' + file.getName().substring(0, file.getName().length() - 6)));
+            }
+        }
+        return classes;
+    }
+
+    public String getAnnotations() throws URISyntaxException, IOException, ClassNotFoundException {
+        this.annotations = "";
+
+        this.getClasses("com.finbourne.lusid.tutorials").forEach(clazz -> {
+            for (Method method : clazz.getMethods())
+            {
+                if (method.isAnnotationPresent(LusidFeature.class))
+                {
+                    this.annotations = this.annotations + method.getAnnotation(LusidFeature.class).value() + "\n";
+                }
+            }
+                }
+        );
+        return this.annotations;
+    }
 
 }
